modify main.h and int_to_string files

+++ main.h
@@ -1,12 +1,12 @@
 #ifndef MAIN_H
 #define MAIN_H
-
+#include <stdint.h>
 #include <stdarg.h>
 int _putchar(char c);
 int _strlen(const char *s, int flag);
-void unsigned_integer_to_string(unsigned int  integer, int base,
+void unsigned_integer_to_string(uint64_t integer, int base,
 	int capitalflag, char *buffer);
-void signed_integer_to_string(int integer, int base,
+void signed_integer_to_string(int64_t integer, int base,
 	int capitalflag, char *buffer);
 void v_printf(const char *format, va_list args);
 void switchFunction (const char *format, va_list args);
